+++ Automation/Settings/EntityPanelSettings.cs
@@ -21,6 +21,7 @@ sealed class EntityPanelSettings : BaseSettings<EntityPanelSettings> {
   const string DescriptionScriptShortLocKey = "IgorZ.Automation.Settings.EntityPanel.RulesDescriptionStyle.ScriptShort";
   const string EvalValuesInConditionsLocKey = "IgorZ.Automation.Settings.EntityPanel.EvalValuesInConditions";
   const string EvalValuesInActionArgumentsLocKey = "IgorZ.Automation.Settings.EntityPanel.EvalValuesInActionArguments";
+  const string MaxRulesShownLocKey = "IgorZ.Automation.Settings.EntityPanel.MaxRulesShown";
 
   protected override string ModId => Configurator.AutomationModId;
 
@@ -54,6 +55,11 @@ sealed class EntityPanelSettings : BaseSettings<EntityPanelSettings> {
   public ModSetting<bool> EvalValuesInActionArgumentsInternal { get; } =
     new(true, ModSettingDescriptor.CreateLocalized(EvalValuesInActionArgumentsLocKey));
 
+  public static int MaxRulesShown { get; private set; }
+
+  public ModSetting<int> MaxRulesShownInternal { get; } =
+    new(6, ModSettingDescriptor.CreateLocalized(MaxRulesShownLocKey));
+
   // ReSharper restore MemberCanBePrivate.Global
   // ReSharper restore InconsistentNaming
   #endregion
@@ -80,5 +86,6 @@ sealed class EntityPanelSettings : BaseSettings<EntityPanelSettings> {
     });
     InstallSettingCallback(EvalValuesInConditionsInternal, v => EvalValuesInConditions = v);
     InstallSettingCallback(EvalValuesInActionArgumentsInternal, v => EvalValuesInActionArguments = v);
+    InstallSettingCallback(MaxRulesShownInternal, v => MaxRulesShown = v);
   }
 }